CBD-4892: Make GRPC more generic to handle SDK testing and metrics
Further iterating the GRPC to make it generic enough to handle SDK
integration testing in future, and generally tidy it up.

+++ performers/java/sdk-performer-java/src/main/java/com/couchbase/sdk/perf/GrpcPerformanceMeasureThread.java
@@ -58,7 +58,7 @@ import java.util.concurrent.atomic.AtomicInteger;
                         (int) (enqueuedFrozen / CHECK_EVERY_X_SECONDS),
                         ignoredFrozen == 0 ? "" : " and ignored " + ignoredFrozen,
                         sentSingleTotal,
-                        sentSingleTotal / totalTimeSecs,
+                        (int) (sentSingleTotal / totalTimeSecs),
                         enqueuedTotal,
                         enqueuedTotal - sentSingleTotal);
